backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
greenlet>=3.0.0

# Document processing
pdfplumber>=0.10.4
python-docx>=1.1.0
tiktoken>=0.8.0

# AI and embeddings
openai>=1.12.0
pinecone>=5.0.1
numpy>=1.26.0

# Configuration
python-dotenv>=1.0.1
pydantic-settings>=2.1.0

# Utilities
aiofiles>=23.2.1
httpx>=0.26.0
tenacity>=8.2.3
